refactor: duplication in src
