+++ src/lib.rs
@@ -2,5 +2,7 @@ pub mod params;
 pub mod rt;
 pub mod split;
 pub mod work;
+pub mod watcher;
+pub mod path;
 
 pub mod spawner;
