+++ src/mspawner.rs
@@ -1,80 +1,12 @@
-use std::io::Write;
-use std::{fs, thread};
-
-use std::time::Duration;
-
 use std::sync::atomic::AtomicUsize;
-use std::sync::atomic::Ordering::Relaxed;
-use std::sync::{mpsc, mpsc::SyncSender, Arc};
-
-use tokio_metrics::RuntimeMonitor;
-
-use std::fs::File;
-use std::path::{Path, PathBuf};
+use std::sync::{mpsc, Arc};
 
 use tokiobench::params;
 use tokiobench::params::metrics as m;
 use tokiobench::rt;
 use tokiobench::spawner;
-
-fn metrics_path() -> PathBuf {
-    let mut path = std::env::current_dir().unwrap();
-
-    path.push("target");
-    path.push("metrics");
-
-    path
-}
-
-fn mk_prefix_dir(folder: &str) -> PathBuf {
-    let mut path = metrics_path();
-    path.push(folder);
-
-    if !Path::exists(&path) {
-        fs::create_dir_all(&path).unwrap();
-    }
-
-    path
-}
-
-fn store(prefix: &Path, name: &str, data: &[u8]) {
-    let result_path = {
-        let mut prefix = PathBuf::from(prefix);
-        prefix.push(name);
-        prefix
-    };
-
-    let mut f = File::create(result_path).unwrap();
-    f.write_all(data).unwrap();
-}
-
-type MetricSyncSender = SyncSender<tokio_metrics::RuntimeMetrics>;
-
-fn run_watcher(
-    metric_tx: MetricSyncSender,
-    rem: Arc<AtomicUsize>,
-    rt_monitor: RuntimeMonitor,
-) -> std::thread::JoinHandle<()> {
-    let thread_handle = thread::spawn(move || {
-        let mut metrics_count = 0;
-
-        for interval in rt_monitor.intervals() {
-            metrics_count += 1;
-            if metrics_count >= m::CHAN_SIZE {
-                panic!("metrics overflow");
-            }
-            metric_tx.send(interval).unwrap();
-
-            if rem.load(Relaxed) == 0 {
-                break;
-            }
-
-            thread::sleep(Duration::from_millis(m::SAMPLE_SLICE));
-        }
-    });
-
-    thread_handle
-}
+use tokiobench::watcher;
+use tokiobench::path::metrics as mpath;
 
 fn run_iter(
     count_down: usize,
@@ -92,7 +24,7 @@ fn run_iter(
         let handle = rt.handle();
         let rt_monitor = tokio_metrics::RuntimeMonitor::new(&handle);
 
-        run_watcher(m_tx, rem, rt_monitor)
+        watcher::run(m_tx, rem, rt_monitor)
     };
 
     rt.block_on(async move {
@@ -108,15 +40,14 @@ fn run_iter(
 
 fn run_metrics(name: &str, count_down: usize, nworkers: usize, bench_fn: spawner::BenchFn) {
     let name = format!("{}_nwork({})", name, nworkers);
-
-    let prefix = mk_prefix_dir(&name);
+    let prefix = mpath::mk_prefix(&name);
 
     for niter in 0..m::N_ITER {
         let metrics = run_iter(count_down, nworkers, bench_fn);
         let metrics_u8 = serde_json::to_vec_pretty(&metrics).unwrap();
 
         let name = format!("iter_{niter}.json");
-        store(&prefix, &name, &metrics_u8);
+        mpath::store(&prefix, &name, &metrics_u8);
     }
 }
